chore: Remove unnecessary casts (#6944)
* Remove unnecessary casts

* Add @SuppressWarning to false positives for UnncessaryCast

See https://github.com/pmd/pmd/issues/6943

* Refactor duplicate code into function

+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/design/StdCyclomaticComplexityRule.java
@@ -74,7 +74,7 @@ public class StdCyclomaticComplexityRule extends AbstractApexRule {
         }
 
         public int getComplexityAverage() {
-            return (double) methodCount == 0 ? 1 : (int) Math.rint((double) decisionPoints / (double) methodCount);
+            return methodCount == 0 ? 1 : (int) Math.rint((double) decisionPoints / methodCount);
         }
     }
